# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.2
components:
    securitySchemes:
        bearerAuth:            # arbitrary name for the security scheme
          type: http
          scheme: bearer
          bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    schemas:
      User:
        title: User
        description: This object represent a single user
        type: object
        properties:
          idUser:
            description: the identifier of the user
            type: integer
            example: 1
            readOnly: true
          username:
            description: the nickname of the user
            type: string
            minLength: 3
            maxLength: 16
            pattern: 'ˆ.+?$'
            example: 'lorenzo23'
      Image:
        title: Image
        description: This object represents a single image (photo)
        type: object
        properties:
          idImage:
            description: the identifier of the image
            type: integer
            example: 1
            readOnly: true
          idOwner:
            description: the identifier of the user who posted the foto
            type: integer
            example: 1
            readOnly: true
          dateTime:
            description: the date and the time the photo was posted
            type: string
            format: date-time
            pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z[-+]\d{2}:\d{2}$'
            example: '2023-11-25T15:30:00Z+03:00'
            minLength: 20
            maxLength: 29
            readOnly: true
          file:
            description: the image file
            type: string
            format: binary
          likesCount:
            description: the number of likes
            type: integer
            example: 10
          commentsCount:
            description: the number of comments
            type: integer
            example: 5
          comments:
            description: the comments
            type: array
            items: { $ref: "#/components/schemas/Comment" }
...
      Comment:
        title: Comment
        description: This object represents a single comment
        type: object
        properties:
          idComment:
            description: the identifier of the comment
            type: integer
            example: 1
            readOnly: true
          idImageCommented:
            description: the identifier of the photo that was commented on
            type: integer
            example: 1
            readOnly: true
          idOwner:
            description: the identifier of the owner of the photo
            type: integer
            example: 1
          idUserWriter:
            description: the identifier of the user who commented on the photo
            type: integer
            example: 1
          text:
            description: the text of the comment
            type: string
            pattern: '.+'
            minLength: 1
            maxLength: 200
            example: beautiful picture!
      UserProfile:
        title: UserProfile
        description: this object represents a profile of an user
        type: object
        properties:
          User:
            $ref: "#/components/schemas/User"
          followCount:
            description: the number of the users followed
            type: integer
            example: 1
          followerCount:
            description: the number of the followers
            type: integer
            example: 1
          Image:
            $ref: "#/components/schemas/Image"
          followStatus:
            description: true if who is using Wasaphoto follows this user, else -> false
            type: boolean
            example: true
      Stream:
        title: Stream
        description: this object represents an user's stream
        type: object
        properties:
          Image:
            $ref: "#/components/schemas/Image"
          Username:
            description: the username of the owner of the image
            type: string
            example: lorenzo
...
paths:
...
  /users/{idUser}/bans/:
    #banUser
    post:
      security:
        - bearerAuth: []
      tags: ["ban"]
      operationId: banUser
      summary: ban an existing user
      description: |
        ban an existing (and not banned) user given its id
      parameters:
      - name: idUser
        in: path
        description: the identifier of the user you want to ban
        schema:
          type: integer
          example: 1
        required: true
      responses:
        "200":
          description: user banned successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: user already banned/you can't ban yourself, not banned
          content:
            application/json:
              schema:
                description: returns the idUserToBan
                type: object
                properties:
                  idUserToBan:
                    description: the id of the user to ban
                    type: integer
                    example: 1
        "404":
          description: user not found, user not banned
          content:
            application/json:
              schema:
                description: returns the idUserToBan
                type: object
                properties:
                  idUserToBan:
                    description: the identifier of the user to ban
                    type: integer
                    example: 1
        "401":
          description: you must be logged in, not banned

    #getBannedList
    get:
      security:
        - bearerAuth: []
      tags: ["ban"]
      operationId: getBannedList
      summary: returns the users who you have banned
      description: |
        get the list of the users who you have banned
      parameters:
      - name: idUser
        in: path
        description: the identifier of the user requesting the list
        schema:
          type: integer
          example: 1
        required: true
      responses:
        "200":
          description: get banned list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403":
          description: you can't see another user's banned list
          content:
            application/json:
              schema:
                description: returns the idUser
                type: object
                properties:
                  idUser:
                    description: the wrong id user
                    type: integer
                    example: 1
        "404":
          description: user not found
          content:
            application/json:
              schema:
                description: returns the idUser
                type: object
                properties:
                  idUser:
                    description: the wrong id user
                    type: integer
                    example: 1
        "401":
          description: you must be logged in, not banned
...
  /users/{idUser}:
    #getUserProfile
    get:
      security:
        - bearerAuth: []
      tags: ["user"]
      operationId: getUserProfile
      summary: return the user profile
      description: |
        return the user profile given its id, the user profile is composed by the user’s photos (in reverse chronological order),
        how many photos have been uploaded,
        and the user’s followers and following
      parameters:
        - name: idUser
          in: path
          description: identifier of the user
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: user profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: you must be logged in
        "403":
          description: you cannot see the profile of a user who has banned you/you banned
          content:
            application/json:
              schema:
                description: return the idUser
                type: object
                properties:
                  idUser:
                    description: the identifier of the user
                    type: integer
                    example: 1
        "404":
          description: user not found
          content:
            application/json:
              schema:
                description: returns the idUser
                type: object
                properties:
                  idUser:
                    description: the identifier of the user
                    type: integer
                    example: 1
    
    #setMyUserName
    put:
      security:
        - bearerAuth: []
      tags: ["user"]
      operationId: setMyUserName
      summary: set/update username
      description: |
        update a user's username given their id and new username, returns all User's informations
      parameters:
      - name: idUser
        in: path
        schema:
          type: integer
          example: 1
        required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newUsername:
                  description: the new username
                  type: string
                  pattern: 'ˆ.+?$'
                  minLength: 3
                  maxLength: 16
                  example: 'username2'
      responses:
        "200":
          description: username successfully changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404":
          description: user id not found, username not changed
          content:
            application/json:
              schema:
                description: returns the idUser
                type: object
                properties:
                  idUser:
                    description: the identifier of the user
                    type: integer
                    example: 1
        "409":
          description: username already used, username not changed
          content:
            application/json:
              schema:
                description: returns the username
                type: object
                properties:
                  username:
                    description: the username choosen
                    type: string
                    example: paper23
        "403":
          description: you can't change the userame of another user, username not changed
          content:
            application/json:
              schema:
                description: returns the idUser and the token
                type: object
                properties:
                  idUser:
                    description: the identifier of the user whose name you want to change
                    type: integer
                    example: 1
                  token:
                    description: the id of the user who is logged into WASAPhoto
                    type: integer
                    example: 2
        "401":
          description: you must be logged in , username not changed
...
  /users/{idUser}/stream:
    #getMyStream
    get:
      security:
        - bearerAuth: []
      tags: ["stream"]
      operationId: getMyStream
      summary: get images of followed users
      description: |
        get images of followed users in chronological order (from the most recent to the oldest)
      parameters:
      - name: idUser
        in: path
        required: true
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: my stream
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 100
                items:
                  $ref: "#/components/schemas/Stream"
        "401":
          description: you must be logged in
        "403":
          description: you cant' see another user's stream
        "404":
          description: user not found
          content:
            application/json:
              schema:
                description: returns the idUser
                type: object
                properties:
                  idUser:
                    description: the identifier of the user
                    type: integer
                    example: 1
  /users/{username}/:
    #searchUser
    get:
      security:
        - bearerAuth: []
      tags: ["user"]
      operationId: searchUser
      summary: search an existing user
      description: |
        searching an existing user given an username
      parameters:
      - name: username
        in: path
        schema:
          type: string
          example: lorenzo
        required: true
      responses:
        "200":
          description: user succesfully finded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404":
          description: user not found
          content:
            application/json:
              schema:
                description: returns the username searched
                type: object
                properties:
                  username:
                    description: username searched
                    type: string
                    example: lorenzo
        "401":
          description: you must be logged in
        "403":
          description: you have been banned from the user and/or you have banned the user/you, or, you can't search yourself
          content:
            application/json:
              schema:
                description: returns an error string and the username searched
                type: object
                properties:
                  username:
                    description: the username searched
                    type: string
                    example: lorenzo